requirements.txt
numpy
opencv-python
pillow
numpy
thread6
tk